Add scream commands and initial screams list
Introduced two new commands to the bot: '/scream' for outputting a
random scream from a predefined list, and '/screamedit' allowing users
to add new screams if they're unique. A JSON file to store the scream
list was also created, along with utility functions for reading from and
writing to this file.

These enhancements aim to enrich the bot's interactive features,
offering more engagement and fun to the users.

v